+++ src/main/java/com/hust/coffeeshop/coffeeshopproject/controller/EventPromotionController.java
@@ -1,4 +1,3 @@
-// src/main/java/com/hust/coffeeshop/coffeeshopproject/controller/EventPromotionController.java
 package com.hust.coffeeshop.coffeeshopproject.controller;
 
 import com.hust.coffeeshop.coffeeshopproject.dto.EventPromotionRequestDTO;
@@ -19,31 +18,27 @@ public class EventPromotionController {
     @Autowired
     private EventPromotionService eventPromotionService;
 
-    // CREATE
     @PostMapping
     public ResponseEntity<EventPromotionResponseDTO> createPromotion(@RequestBody EventPromotionRequestDTO requestDTO) {
         EventPromotionResponseDTO createdPromotion = eventPromotionService.createPromotion(requestDTO);
         return new ResponseEntity<>(createdPromotion, HttpStatus.CREATED);
     }
 
-    // READ ALL
     @GetMapping
     public ResponseEntity<List<EventPromotionResponseDTO>> getAllPromotions() {
         List<EventPromotionResponseDTO> promotions = eventPromotionService.getAllPromotions();
         return new ResponseEntity<>(promotions, HttpStatus.OK);
     }
 
-    // READ BY ID
     @GetMapping("/{id}")
-    public ResponseEntity<EventPromotionResponseDTO> getPromotionById(@PathVariable Integer id) {
+    public ResponseEntity<EventPromotionResponseDTO> getPromotionById(@PathVariable Long id) { // SỬA TẠI ĐÂY
         return eventPromotionService.getPromotionById(id)
                 .map(promotionDTO -> new ResponseEntity<>(promotionDTO, HttpStatus.OK))
                 .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    // UPDATE
     @PutMapping("/{id}")
-    public ResponseEntity<EventPromotionResponseDTO> updatePromotion(@PathVariable Integer id, @RequestBody EventPromotionRequestDTO requestDTO) {
+    public ResponseEntity<EventPromotionResponseDTO> updatePromotion(@PathVariable Long id, @RequestBody EventPromotionRequestDTO requestDTO) { // SỬA TẠI ĐÂY
         try {
             EventPromotionResponseDTO updatedPromotion = eventPromotionService.updatePromotion(id, requestDTO);
             return new ResponseEntity<>(updatedPromotion, HttpStatus.OK);
@@ -52,12 +47,11 @@ public class EventPromotionController {
         }
     }
 
-    // DELETE
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePromotion(@PathVariable Integer id) {
+    public ResponseEntity<Void> deletePromotion(@PathVariable Long id) { // SỬA TẠI ĐÂY
         try {
             eventPromotionService.deletePromotion(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // 204 No Content
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (EntityNotFoundException e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
